Introduced open64 and resync64; made test program for old DM interface

+++ tables/Tables/BaseTable.cc
@@ -618,7 +618,7 @@ void BaseTable::removeRow (const Vector<rownr_t>& rownrs)
     Vector<rownr_t> rownrsCopy;
     rownrsCopy = rownrs;
     genSort (rownrsCopy);
-    for (Int i=rownrsCopy.nelements()-1; i>=0; i--) {
+    for (Int64 i=rownrsCopy.nelements()-1; i>=0; i--) {
 	removeRow (rownrsCopy(i));
     }
 }
